Add feature to evaluate all legal combinations of gov_pmt, crop_ins
To implement this feature, the ability to pass separate override dicts to the CashFlow constructor was added.  The duplicated validation in Indemnity was removed since Indemnities are almosed always used in the context of CropIns.  Also, it turns out that ECO cannot be placed on a Area stack, although unlike SCO, it can be used with an ARC-CO program.