Make crop-specific and premium attributes dicts, use ints for crops
Many of the attributes imported from files have names that differ only by the crop.
Changing to dicts with key crop reduces the number of attributes significantly.For premiums, we can reduce the number of properties to just
three, one for the base premium and one for each option type.  In this case the dict key is
a tuple of 4 or 3 ints.  This should simplify the step of replacing the text files with
database tables.